refactor(firebase): move firebase setup to another file
Move firebase setup to another file, to allow use firestore in multiple files.
